requirements.txt
albumentations==1.1.0
easyocr==1.4.1
matplotlib==3.5.0
numpy==1.21.4
opencv_python_headless==4.5.4.60
opencv-python~=4.5.5.62
Pillow~=8.2.0
pyocr==0.8.1
torch==1.10.1
